instantiate headers as set to prevent dupes; tests
- rename for more sane test discovery
- mv 'live tests' to another file
- be more careful about using default headers vs 'incl_headers'. There were some edge cases that would fail. They don't fail now, and the behavior is more explicit.
- add tests for set-based headers to prove all sorts of variations.
